Progress continues. Index match is working properly. I have begun the process of updating spreadsheets. Updating the easy spreadsheets looks like is working functionally, I just need the code to add a tab and input the account value from the Dict.
More effort will need to be put into understanding how to update the accounts that had no activity as well as the ability to create new spreadsheets if there is a account in the GL that is not in the recon directory. Note: there will need to be some prompt that asks "Does this account have a document to verify against?"

Perhapse I will need to incorporate a database that memorizes what accounts can be compared against statements and which accounts are (contra accounts?) nothing to reconcile against.