The spreadsheets no longer need to be "formatted" because there is a "move and copy" funcion in openpyxl that allows me to create a new tab based on the previous tab.
As such, the spreadsheets can be updated easily now. The main python file will need to iterate through every recon name (.xlsx) and fill in the apropriate account balance in J4.

Still need to figure out try/excepts for if the account balance hasn't cahnged or if a new recon needs to be created.