The program now loops through every spreadsheet within a directory and if my Dictionary has a key, it will put in the correct account balance.
Still need to work on creating new workbooks when an account is added through quickbooks without having a recon created as well.

For accounts that have no activity, the a "Key Error" will allow a move/copy tab to happen but will prevent the new account balance from being updated.